refactor: move common video schemas to dedicated module and update imports
