+++ src/schema/videos/common.ts
@@ -1,0 +1,28 @@
+import { z } from 'zod';
+import { validateMediaURL } from 'src/services/videos/convert/validator';
+
+const videoUrlSchema = z
+  .string()
+  .url()
+  .regex(/^https:\/\//i, 'URL must use HTTPS')
+  .refine(
+    url => {
+      const result = validateMediaURL(url);
+      return result.platform !== null || result.fileType !== null;
+    },
+    {
+      message: 'URL must be a valid media file or from a supported platform',
+    }
+  );
+
+/**
+ * Hasura event trigger schema
+ */
+
+const EventMetadataSchema = z.object({
+  id: z.string(),
+  span_id: z.string(),
+  trace_id: z.string(),
+});
+
+export { videoUrlSchema, EventMetadataSchema };
